% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{create_level_results}
\alias{create_level_results}
\title{create_level_results}
\usage{
create_level_results(df)
}
\arguments{
\item{df}{the current data frame used is 'program_data_21'}
}
\value{
a date frame
}
\description{
takes input from the result of create pre_program_dat and returns outcome by levels
}
