% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{create_levels}
\alias{create_levels}
\title{create_stats}
\usage{
create_levels(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
a data frame
}
\description{
function to get numeric variables and stats from pre survey
}
\examples{
df_1 <- create_levels(df)
}
