% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{create_program}
\alias{create_program}
\title{create_program}
\usage{
create_program(df)
}
\arguments{
\item{df}{output from read_raw}
}
\value{
a data frame
}
\description{
initial clean upof data set from create_raw
}
