% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{create_program_dat}
\alias{create_program_dat}
\title{create_program_dat}
\usage{
create_program_dat(df, time)
}
\arguments{
\item{df}{that data frame that is used.}

\item{time}{consists of pre/post.}
}
\value{
a data frame with just pre surveys
}
\description{
create data frame with just pre surveys and wipe out duplicate girl_codes from df
}
