% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{create_raw}
\alias{create_raw}
\title{create_raw}
\usage{
create_raw(start_date)
}
\arguments{
\item{start_date}{Character string to tell Qualtrics to pull data from this date}
}
\value{
The return value, if any, from executing the utility.
}
\description{
A wraper for the fetch_survey function from qualtRics
}
