% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{create_summary}
\alias{create_summary}
\title{create summary}
\usage{
create_summary(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
df
}
\description{
selects and summarizes mean scores and display in df.
}
\examples{
 create_summary(df)
}
