% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sites.R
\name{grant_g}
\alias{grant_g}
\title{grant_g}
\usage{
grant_g(df)
}
\arguments{
\item{df}{is the data frame and it should always be grants}
}
\value{
a ggplot
}
\description{
A ggplot graph of valid survey count
}
