% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_theme.R
\name{gs_theme_538}
\alias{gs_theme_538}
\title{gs_theme_538}
\usage{
gs_theme_538(data, ...)
}
\arguments{
\item{data}{the data frame you want to display as gt object.}

\item{...}{extra param.}
}
\value{
a gt
}
\description{
gs_theme_538
}
