% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_theme.R
\name{outreach_palette}
\alias{outreach_palette}
\title{outreach_palette}
\usage{
outreach_palette(name, n)
}
\arguments{
\item{name}{of the palette}

\item{n}{is the number of colors}
}
\value{
a color scheme
}
\description{
outreach_palette
}
