% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_theme.R
\docType{data}
\name{outreach_palettes}
\alias{outreach_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 1.
}
\usage{
outreach_palettes
}
\description{
Use \code{\link{gs_palette}} to construct palettes of desired length.
}
\keyword{datasets}
