% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sites.R
\name{program_sites}
\alias{program_sites}
\title{Program Sites}
\usage{
program_sites(df)
}
\arguments{
\item{df}{Should alawys be survy data}
}
\value{
a data frame
}
\description{
Function to parse program sites
}
