% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sites.R
\name{report_site}
\alias{report_site}
\title{Site Report}
\usage{
report_site(df)
}
\arguments{
\item{df}{Always use program sites}
}
\value{
A table from the janitor package
}
\description{
Function to put sites in a table
}
