% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sites.R
\name{top_sites_tbl}
\alias{top_sites_tbl}
\title{Top Sites Report}
\usage{
top_sites_tbl(df)
}
\arguments{
\item{df}{Should be df result of program_sites}
}
\value{
A tbl
}
\description{
Report to display the top 15 sites
}
